scraper/requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
